FIFO Principle in Accounting
Requirement:
FIFO Principle in Accounting
In any financial transaction there may be profit or loss associated
with that transaction. The party that receives the proceeds from such a
transaction needs to calculate the gains or losses realized in the
transaction. For example as if you buy an asset (e.g., house, car,
laptop, or shares) you need to record the cost of the asset and when
you sell the same you need to calculate your gains by subtracting the
cost from the sale proceeds (amount received from the sale). However,
if you buy and sell frequently, you need to follow the FIFO principle,
that is, if you have purchased some assets (or items) of the same type
on several different dates- when you sell that asset (or item)- the
oldest asset of your collection is considered to be sold and the cost
of that needs to be used for calculating the gains.
So when you are buying and selling stocks in different quantities, an
inventory record needs to be kept that stores the purchase transaction
with the name of the asset (or item) along with quantity, the date and
cost (corresponding to each item). When you carry out a “sell”
transaction, you need to calculate the cost basis for the assets sold
as per the record.
The same program essentially can be used for tracking inventory of any
other items with costs basis for any other business operations.
Assignment
For this assignment write a program that keeps record of all
chronological “ins” (purchases) and “outs” (sale or consumption) on a
FIFO basis. Use simple serial numbers to indicate oldest to newest.
Input
The program should read from a file (pa1input.txt) with at least 25
entries of transactions of different items in the format:
“in x1: q1 at $p1 each” or “out x2: q2 at $s2 each”
where qi is the transaction quantity, pi is the purchase price and si
is the selling price of each the item xi .
For using this program for general inventory, the second input may be
worded as:
“out x2: q2 at $0 each”.
Output
Output should be written to an output file (pa1output.txt).
If at any time out quantity exceeds the inventory, it should give an
appropriate error message and stop the program.
Otherwise, in the output, you should have two tables- one for the final
(current) inventory (showing all remaining “in” entries,) and one for
the cost basis for each of the “out” transaction in the following
format:
Table 1: Current Inventory Serial # Item Quantity Cost # xi qi pi ----
----
Table 2: Cost Basis and Gains
Serial # Item Quantity Cost Basis ($) Gains ($)
# xi qi 􏰅(qi.pi) 􏰅qi.si ) – (􏰅qi.pi)
---- ----
- -
Table 2 may look like the following if the program is used for
maintaining general inventory.
# xi qi 􏰅(qi.pi) – (􏰅qi.pi)
---- ----
- -